data-ingestion-dashboard/requirements.txt

pyspark
kafka-python
delta-spark
pandas
matplotlib
tensorflow
joblib
scikit-learn
streamlit